pyproject.toml
[tool.poetry]
name = "ai-exp-learning"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
gym = "0.22.0"
pyworld = "^0.3.4"
pysptk = "^1.0.1"
dt = "^1.1.65"
pyyaml = "^6.0.2"
numpy = "1.26.0"
torchvision = "0.14.1"
pygame = "^2.6.1"
six = "^1.16.0"
matplotlib = "^3.9.2"
dtw = "^1.4.0"
ultralytics = "^8.3.15"
label-studio = "^1.13.1"
moviepy = "^1.0.3"
scikit-learn = "^1.5.2"
wandb = "^0.18.5"
opencv-python = "^4.10.0.84"
transformers = "^4.46.0"
scikit-optimize = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"